feat:moving profile compoents moving profile componets to @/modules/ProfileModule/components
