+++ frontend/tmp.env
@@ -1,1 +1,1 @@
-REACT_APP_BACKEND_SERVER="https://your-remote-dev-server.com"
+REACT_APP_BACKEND_SERVER="https://dev-server.idurarapp.com"
